+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 FEATURES:
 * Azure China support (China East, China East 2, China North, China North 2, and China North 3)
 
+BUG FIXES:
+* fix: brownfield cc_lb defaults for non-zpa private dns deployments
+* 
 ENHANCEMENTS:
 * ZSEC bash script support for Azure China regions