+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 ## TBD (UNRELEASED)
+
+FEATURES:
+* Azure China support (China East, China East 2, China North, China North 2, and China North 3)
+
+BUG FIXES:
+* fix: brownfield cc_lb defaults for non-zpa private dns deployments
+* 
 ENHANCEMENTS:
-* add: variable support_access_enabled for dynamic NSG rule creation for Zscaler Support Tunnel access
+* ZSEC bash script support for Azure China regions
+* * add: variable support_access_enabled for dynamic NSG rule creation for Zscaler Support Tunnel access
 * add: zsec prompt for support tunnel rule creation
 
 ## v0.3.0 (September 30, 2023)